# payments-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Payments Reliability API",
    "description": "Production-grade webhook processing service for Stripe and PayPal",
    "version": "0.2.0-rc"
  },
  "paths": {
...
    "/reconciliation/summary": {
      "get": {
        "summary": "Reconciliation Summary",
        "description": "Get reconciliation summary for the specified time window.\n\nArgs:\n    hours_back: Number of hours to look back (1-168)\n\n    Returns:\n        Reconciliation summary with event counts and anomalies",
        "operationId": "reconciliation_summary_reconciliation_summary_get",
        "parameters": [
          {
            "name": "hours_back",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 1,
              "default": 24,
              "title": "Hours Back"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/events": {
      "get": {
        "summary": "List Events",
        "description": "List recent payment events for debugging.",
        "operationId": "list_events_admin_events_get",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReconciliationSummary": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date-time",
            "title": "Period Start",
            "description": "Summary period start"
          },
          "period_end": {
            "type": "string",
            "format": "date-time",
            "title": "Period End",
            "description": "Summary period end"
          },
          "total_events": {
            "type": "integer",
            "title": "Total Events",
            "description": "Total events processed"
          },
          "events_by_provider": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Events By Provider",
            "description": "Events grouped by provider"
          },
          "events_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Events By Type",
            "description": "Events grouped by type"
          },
          "anomalies": {
            "items": {},
            "type": "array",
            "title": "Anomalies",
            "description": "Detected anomalies"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At",
            "description": "Summary generation time"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_end",
          "total_events",
          "events_by_provider",
          "events_by_type",
          "anomalies",
          "generated_at"
        ],
        "title": "ReconciliationSummary",
        "description": "Reconciliation summary response."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}